requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.11
emoji==1.6.3
firebase-admin==5.2.0
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.2.1
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.4
idna==3.3
msgpack==1.0.3
multidict==6.0.2
packaging==21.3
proto-plus==1.20.0
protobuf==3.19.4
py-cord==1.7.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.16.0
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
yarl==1.7.2
